Updated to one loop

+++ internal/theme/hyprland.go
@@ -11,24 +11,24 @@ type Hyprland struct {
 	PrimaryFont string `json:"primaryFont"`
 	SecondaryFont string `json:"secondaryFont"`
 
-	GapsIn int8 `json:"gapsIn"`
-	GapsOut int8 `json:"gapsOut"`
-	BorderSize int8 `json:"borderSize"`
+	GapsIn int `json:"gapsIn"`
+	GapsOut int `json:"gapsOut"`
+	BorderSize int `json:"borderSize"`
 	ColActiveBorder1 string `json:"colActiveBorder1"`
 	ColActiveBorder2 string `json:"colActiveBorder2"`
 	ColActiveBorderDeg string `json:"colActiveBorderDeg"`
 	ColInactiveBorder string `json:"colInactiveBorder"`
 
 	Blur bool `json:"blur"`
-	BlurSize int8 `json:"blurSize"`
-	BlurPasses int8 `json:"blurPasses"`
-	Rounding int8 `json:"rounding"`
+	BlurSize int `json:"blurSize"`
+	BlurPasses int `json:"blurPasses"`
+	Rounding int `json:"rounding"`
 	Shadow bool `json:"shadow"`
-	ShadowRange int8 `json:"shadowRange"`
-	ShadowRenderPower int8 `json:"shadowRenderPower"`
+	ShadowRange int `json:"shadowRange"`
+	ShadowRenderPower int `json:"shadowRenderPower"`
 	ShadowColor string `json:"shadowColor"`
 
-	FootOpacity float32 `json:"footOpacity"`
+	FootOpacity float64 `json:"footOpacity"`
 }
 
 func (tool Hyprland) Render(templatePath, outputPath string) error {
